Add some things 'n stuffs

+++ setup.sh
@@ -6,4 +6,9 @@ if [ ! -d "$HOME/.config/nvim" ]; then
   mkdir $HOME/.config/nvim
 fi
 
-ln -s ${dir}/.config/nvim/init.lua $HOME/.config/nvim/init.lua
+if [ -L $HOME/.config/nvim/init.lua ] && [ -f $HOME/.config/nvim/init.lua ]; then
+  rm -f $HOME/.config/nvim/init.lua
+  ln -s ${dir}/.config/nvim/init.lua $HOME/.config/nvim/init.lua
+else
+  ln -s ${dir}/.config/nvim/init.lua $HOME/.config/nvim/init.lua
+fi
